add more tests, add input validation to all ui components

+++ analytica/src/main/java/analytica/domain/Regression.java
@@ -65,7 +65,7 @@ public class Regression {
         }                
     }
     
-    public long getNumberOfValuePairs() {
+    public int getNumberOfValuePairs() {
         return this.index;
     }
